Add HTTP Range support for streamed static files
Building on the earlier streamResponse work: normalizeHeaders() now
parses Range (all three bytes= forms - start-end, start-, -suffix),
validates it against the known file size, and switches the status
line to 206 Partial Content / 416 Range Not Satisfiable accordingly -
416 also clears streamResponse so ClientHandler never tries to send a
body. streamResponseTo() seeks via FileChannel.position() instead of
the unreliable InputStream.skip() and reads exactly the requested
span, treating a premature EOF as a real error rather than a normal
stream end.

msg now reflects the outcome (plain stream / 206 with the exact
byte range / 416 with the rejected Range and file size) instead of
always reading the same "200 OK" line, so the log can tell a cached
hit apart from a disk-streamed response. Client mid-stream aborts
(routine during video seeking - Chrome drops the stale Range request
and issues a new one) are now caught right at streamResponseTo() and
reported as "перервано клієнтом" instead of disappearing silently
into the generic IOException handler.

Widened the size type from int to long the whole way through -
FileCacheManager.isStreamResponse() -> GetRes/GetRes80 -> both
HTTPResponse constructors -> streamResponseLength/byteBegin/byteEnd -
so a >2GB file wouldn't silently misreport its length.

Added video/mp4 to the Content-Type whitelist (was 415 before) and
load-tested against three real local video files (up to 355MB) using
<video> seek/scrub in Chrome: correct 206 responses, accurate
Content-Range, no worker crashes, no leaked sockets on abort.

+++ servak_na_linux_2_java/source/GetRes.java
@@ -74,7 +74,7 @@ public final class GetRes {
 			if (!filePath.toFile().isDirectory()) {
 				if (filePath.toFile().exists()) {
 					try {
-						int isStreamResponse = FileCacheManager.isStreamResponse(filePath.toString());
+						long isStreamResponse = FileCacheManager.isStreamResponse(filePath.toString());
 						if (isStreamResponse > 0) {
 							if (httpRequest.method.compareTo("HEAD") == 0)
 								return new HTTPResponse(FileCacheManager.getFileStreamResponse(filePath.toString()), requestedFile, isStreamResponse, true);
