+++ servak_na_linux_2_java/source/GetRes80.java
@@ -29,7 +29,7 @@ public final class GetRes80 {
 		if (filePath.startsWith(Paths.get(Configs.getParam("www80_directory")))) {
 			if (filePath.toFile().exists() && !filePath.toFile().isDirectory()) {
 				try {
-					int isStreamResponse = FileCacheManager.isStreamResponse(filePath.toString());
+					long isStreamResponse = FileCacheManager.isStreamResponse(filePath.toString());
 					if (isStreamResponse > 0) {
 						if (httpRequest.method.compareTo("HEAD") == 0)
 							return new HTTPResponse(FileCacheManager.getFileStreamResponse(filePath.toString()), requestedFile, isStreamResponse, true);
